refactor channel page structure
