+++ internal/admin/controller/channel/handler.go
@@ -326,6 +326,9 @@ func AddChannel(c *gin.Context) {
 	c.JSON(http.StatusOK, gin.H{
 		"success": true,
 		"message": "",
+		"data": gin.H{
+			"id": channel.Id,
+		},
 	})
 	return
 }
